Move controller functions from server to separate file
