+++ .env.example
@@ -3,3 +3,4 @@ DB_HOST=YOUR_HOST
 DB_DATABASE=YOUR_DATABASE_NAME
 DB_PASSWORD=YOUR_PASSWORD
 DB_PORT=YOUR_PORT_NUMBER
+PORT=SERVER_PORT
